+++ cmd/lino/main.go
@@ -33,6 +33,7 @@ import (
 	"github.com/cgi-fr/lino/internal/app/id"
 	"github.com/cgi-fr/lino/internal/app/pull"
 	"github.com/cgi-fr/lino/internal/app/push"
+	"github.com/cgi-fr/lino/internal/app/query"
 	"github.com/cgi-fr/lino/internal/app/relation"
 	"github.com/cgi-fr/lino/internal/app/sequence"
 	"github.com/cgi-fr/lino/internal/app/table"
@@ -160,6 +161,7 @@ func init() {
 	rootCmd.AddCommand(push.NewCommand("lino", os.Stderr, os.Stdout, os.Stdin))
 	rootCmd.AddCommand(http.NewCommand("lino", os.Stderr, os.Stdout, os.Stdin))
 	rootCmd.AddCommand(analyse.NewCommand("lino", os.Stderr, os.Stdout, os.Stdin))
+	rootCmd.AddCommand(query.NewCommand("lino", os.Stderr, os.Stdout, os.Stdin))
 }
 
 func initConfig() {
@@ -212,6 +214,7 @@ func initConfig() {
 	id.Inject(idStorageFile, relationStorage(), idExporter(), idJSONStorage(*os.Stdout))
 	pull.Inject(dataconnectorStorage(), relationStorage(), tableStorage(), idStorageFactory(), pullDataSourceFactory(), pullRowExporterFactory(), pullRowReaderFactory(), pullKeyStoreFactory(), traceListner(os.Stderr))
 	push.Inject(dataconnectorStorage(), relationStorage(), tableStorage(), idStorageFactory(), pushDataDestinationFactory(), pushRowIteratorFactory(), pushRowExporterFactory(), pushTranslator(), pushObserver())
+	query.Inject(dataconnectorStorage(), queryDataSourceFactory())
 }
 
 func writeMetricsToFile(statsFile string, statsByte []byte) {